refactor find lemma occurences logic in SentenceServiceImpl

+++ src/main/java/com/example/stanfordnlpgerman/models/dtos/sentence/LemmaOccurenceInSentencesDTO.java
@@ -6,5 +6,5 @@ public interface LemmaOccurenceInSentencesDTO {
 
   String getLemmaText();
 
-  Long getLemmaOccurence();
+  Integer getLemmaOccurence();
 }
